Updated all routes and logins, and added error handling.

+++ models/artifacts.js
@@ -13,7 +13,7 @@ module.exports = function(sequelize, DataTypes) {
       allowNull: false
     },
     price: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.FLOAT(11, 2),
       allowNull: false
     }
   });
